Moved Leaderboard class to types.py
